requirements.txt
curses
figlet
